fix(xrpl): uint32 ledger entry type fields and transaction fields

+++ xrpl/ledger-entry-types/signer_list.go
@@ -12,11 +12,11 @@ type SignerList struct {
 	LedgerEntryType   EntryType
 	Flags             SignerListFlags
 	PreviousTxnID     string
-	PreviousTxnLgrSeq uint64
+	PreviousTxnLgrSeq uint32
 	OwnerNode         string
 	SignerEntries     []SignerEntryWrapper
-	SignerListID      uint64
-	SignerQuorum      uint64
+	SignerListID      uint32
+	SignerQuorum      uint32
 }
 
 type SignerEntryWrapper struct {
@@ -25,7 +25,7 @@ type SignerEntryWrapper struct {
 
 type SignerEntry struct {
 	Account       types.Address
-	SignerWeight  uint64
+	SignerWeight  uint16
 	WalletLocator types.Hash256 `json:",omitempty"`
 }
